requirements.txt
fastapi
uvicorn
python-multipart
pydantic-settings
PyMuPDF
sentence-transformers
faiss-cpu
httpx
pydantic
python-docx
